requirements.txt
numpy==1.23.5
pandas==2.0.2
Pillow==9.5.0
Pillow==9.5.0
streamlit==1.23.1
tensorflow==2.12.0
tensorflow_datasets==4.9.2
tensorflow_intel==2.12.0
